backend/requirements.txt
aiosqlite
argon2-cffi
asyncpg
boto3
email-validator
fastapi
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
httpx
passlib
psycopg
pydantic
pydantic-settings
pyjwt
python-dotenv
python-multipart
resend
slowapi
sqlalchemy
uvicorn